+++ Changelog.md
@@ -4,3 +4,5 @@ Changelog
  * (20/3/2013) Added Wordpress to the solutions.
  
  * (20/3/2013) Updated Wordpress to 3.5.1.
+ 
+ * (20/3/2013) Added Nodeserver to the solutions.